Remove intent receiver foreground and background, use pwa browser fragment timing instead
